#15 - Rename: SearchHeader 파일명 수정
