+++ src/pages/SearchPage.jsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react';
 import SearchContainer from '../components/search/SearchContainer';
-import SearchHeder from '../components/search/SearchHeder';
+import SearchHeader from '../components/search/SearchHeader';
 
 function SearchPage() {
 	const [searchList, setSearchList] = useState(null);
 
 	return (
 		<>
-			<SearchHeder setSearchList={setSearchList} />
+			<SearchHeader setSearchList={setSearchList} />
 			<SearchContainer searchList={searchList} />
 		</>
 	);
